Sending player using Byte Stream
This example uses a Byte Stream to send a player from the Client to
the Server. This takes 188 bytes which is wasteful.
Homework: Write a method which sends the same player using a Bit stream
and the compression techniques we have discussed. Send me a link to a
Github repo and the number of bits you needed to send the player